Use the CA bundle from the credentials when possible
oVirt API over SSL should be verified against a proper CA. It is
expected by the deployment to pass the CA bundle that container oVirt's
CA and keep it in ovirt-credentials secret, under 'ovirt_ca_bundle'.

When it's there, save it to disk, and use it as the api connection CA
file.

Another option was to seriailize the secret and extract the CA bundle
directly to /etc/ssl/certs but that mandates a deployment change, and
moving that logic to the deployment part.

If the CA bundle doesn't exist the deploymnent can still handle that
by changing the CA sources on the container.

Signed-off-by: Roy Golan <rgolan@redhat.com>