Reconcile network addresses according to VM status
MOTIVATION
This fix handles the network addresses reconciliation problems which leads to
mishandling of status and phase reporting and lack of linkage, which causes the
machine to apear in the wrong state.

The main issue is around network reconciliation, and for that a major refactoring been made
to better address this. See [1] below.

RESULT
While the VM is not up it still doesn't have any IP addresses and this should be reported as nil addresses.
If the VM is UP it means we have to report error, and that will trigger a future reconcilliation.
A change in the node object that, that happen when a VM goes down for example, triggers reconciliation
and that should change the list of addresses and state accordingly. When the VM boots a reverse thing
happens - the machine boots with no IP addresses, till the node object is detecting a change because
kubelet starts responding. That triggers a reconcillation that will detect the addresses correctly. If
not that should keep firing reconcillation till it does.

MODIFICATION
[1] structural changes:
Instead of having pile of changes in one or two methods, every change to
the status or spec has its own function.
The high level break down of the any handling now is:
- create or delete the vm. no-op on the vm for update
- reconcile provider id
- reconcile network addresses
- reconcile annotations
- reconcile providerStatus
- update machine
- update machine/status

Signed-off-by: Roy Golan <rgolan@redhat.com>