Remove machinesClient, osClient, kubeClient from Actuator
This commit removes the machinesClient and osClient from the machine scope and actuator code.
Instead we will use the kube client to patch the machine object and get the infra info.
This change will help us with testing the machine scope since it is one client that we don't need to mock plus it will align us with other providers